Added documentation, remove redundant code

+++ lack_of_oxygen/lack_of_oxygen/Utility/Constant.h
@@ -240,7 +240,7 @@ namespace lof {
 
 	// ------------------------------ Player_Script.cpp --------------------------------
 
-	constexpr const unsigned int DEFAULT_LR_FORCE_MAG = 10000.0f; 
+	constexpr const float DEFAULT_LR_FORCE_MAG = 10000.0f; 
 
 	// ------------------------------ Mining_Script.cpp --------------------------------
 	constexpr const float MINING_COOLDOWN_TIMER = 0.2f;
